examples/agent-runtime/agentic-frameworks/google-adk/pyproject.toml
[project]
name = "google-adk-agent-example"
version = "1.0.0"
description = "Google ADK Agent integration with Novita Agent Runtime example project"
authors = [
    {name = "Jason", email = "jason.zhang@novita.ai"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "novita-sandbox>=1.1.0b1",
    "google-adk",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

